requirements.txt
matplotlib
keras -U
tensorflow_datasets
tensorflow_transform